* EDIT: STL containers and extensions refactoring * DEL: dead code in r2_R_render.cpp & r3_R_render.cpp & r4_R_render.cpp

+++ code/engine.vc2008/xrRender/xrRender/r__dsgraph_render_lods.cpp
@@ -26,22 +26,17 @@ void R_dsgraph_structure::r_dsgraph_render_lods	(bool _setup_zb, bool _clear)
 	float			ssaRange		= r_ssaLOD_A - r_ssaLOD_B;
 	if (ssaRange<EPS_S)	ssaRange	= EPS_S;
 
-	const u32	uiVertexPerImposter	= 4;
-	const u32	uiImpostersFit		= RCache.Vertex.GetSize()
-		/ (firstV->geom->vb_stride*uiVertexPerImposter);
+	const u32		uiVertexPerImposter	= 4;
+	const size_t	uiImpostersFit		= RCache.Vertex.GetSize() / (firstV->geom->vb_stride*uiVertexPerImposter);
 
-	//Msg						("dbg_lods: shid[%d],firstV[%X]",shid,u32((void*)firstV));
-	//Msg						("dbg_lods: shader[%X]",u32((void*)firstV->shader._get()));
-	//Msg						("dbg_lods: shader_E[%X]",u32((void*)cur_S._get()));
-
-	for (u32 i=0; i<lstLODs.size(); i++)
+	for (size_t i = 0; i < lstLODs.size(); i++)
 	{
-		const u32	iBatchSize	= std::min( lstLODs.size()-i, uiImpostersFit);
+		const size_t	iBatchSize	= std::min( lstLODs.size()-i, uiImpostersFit);
 		int			cur_count	= 0;
 		u32			vOffset;
 		FLOD::_hw*	V			= (FLOD::_hw*)RCache.Vertex.Lock	(iBatchSize*uiVertexPerImposter,firstV->geom->vb_stride, vOffset);
 
-		for ( u32 j=0; j<iBatchSize; ++j, ++i )
+		for (size_t j = 0; j < iBatchSize; ++j, ++i)
 		{
 			// sort out redundancy
 			R_dsgraph::_LodItem		&P		= lstLODs[i];
